CI: Add lfs disk info on FAT32 test
1/ Refactor lfs disk info test
2/ Add test on FAT32 to the regime
